+++ ARQSOFT_Spreadsheet_Riera_Wasmer/src/main/java/edu/upc/etsetb/arqsoft/spreadsheet/entities/formulas/EvaluatorPostFix.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public class EvaluatorPostFix {
 
-    public float evaluatePostFix(List<FormulaComponent> formulaComponents) {
+    public double evaluatePostFix(List<FormulaComponent> formulaComponents) {
         VisitorFormulaComponents visitor = new VisitorFormulaComponents();
         
         for(FormulaComponent component : formulaComponents){
