+++ ARQSOFT_Spreadsheet_Riera_Wasmer/src/main/java/edu/upc/etsetb/arqsoft/spreadsheet/entities/formulas/OperatorImpl.java
@@ -37,7 +37,7 @@ public class OperatorImpl implements Operator {
     }
 
     @Override
-    public float getFormulaComponentValue() {
+    public double getFormulaComponentValue() {
         throw new UnsupportedOperationException("OperatorImpl Exception"); 
     }
